Flytt API-parametertypar ut av queries.ts
- StønadskontoParams → api/getStønadskontoParams.ts
- AnnenPartVedtakParams → utils/annenForelderUtils.ts
- DokumentereMorsArbeidParams → types/DokumentereMorsArbeidParams.ts
